Implement drink choice
Implement drink choice that allows the user to input what they are currently drinking. if they say yes, it will add a text box to allow them to say what. if they choose no then it will suggest something to drink. there is a bug here that creates too many text boxes if the user goes back and forth on the selection. will fix later.